refactor transactionsTable into own component
